Don’t break reactivity when storing DOM elements
Depending on the situation and type of element you are storing the browser may throw the following error: “Failed to set an indexed property on '{type}': Accessor properties are not allowed.”

This is because Vue Fn API tried to add accessor properties to certain HTML element types which is not allowed.

In more obscure cases, if this assignment happens when handling an event, Vue will actually break during its patching process where a vnode’s elm propeorty ends up undefined.