Update README
Without stating that these imports need to be used in the main.js file, users might struggle to know where they go. It took me a while of experimentation and looking around to figure it out. It might be nice to show the difference from a vue 2 file to this new change for an example as well, rather than just saying install.